chore: refactor pricing calculator
