+++ app/pricing/pricingv1/components/PricingCalculator/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './PricingCalculatorView'
+export type { Section, PricingCalculatorProps } from './types'
