+++ app/pricing/pricingv1/components/PricingCalculator/types.ts
@@ -1,0 +1,8 @@
+export type Section = 'traces' | 'logs' | 'metrics'
+
+export interface PricingCalculatorProps {
+  show?: Section[]
+  showHeader?: boolean
+  showFooter?: boolean
+  embedded?: boolean
+}
